feat:size_t ap

+++ include/fakeLisp/fklni.h
@@ -6,19 +6,19 @@
 extern "C"{
 #endif
 
-FklVMvalue* fklNiGetArg(uint32_t* ap,FklVMstack*);
-FklVMvalue* fklNiPopTop(uint32_t* ap,FklVMstack*);
-void fklNiReturn(FklVMvalue*,uint32_t* ap,FklVMstack*);
+FklVMvalue* fklNiGetArg(size_t* ap,FklVMstack*);
+FklVMvalue* fklNiPopTop(size_t* ap,FklVMstack*);
+void fklNiReturn(FklVMvalue*,size_t* ap,FklVMstack*);
 void fklNiSetBp(uint64_t nbp,FklVMstack* s);
-int fklNiResBp(uint32_t* ap,FklVMstack*);
+int fklNiResBp(size_t* ap,FklVMstack*);
 void fklNiSetTp(FklVMstack*);
 void fklNiResTp(FklVMstack*);
 void fklNiPopTp(FklVMstack*);
-void fklNiEnd(uint32_t* ap,FklVMstack*);
-void fklNiBegin(uint32_t* ap,FklVMstack*);
+void fklNiEnd(size_t* ap,FklVMstack*);
+void fklNiBegin(size_t* ap,FklVMstack*);
 FklVMvalue** fklNiGetTopSlot(FklVMstack*);
 #define FKL_NI_BEGIN(EXE) FklVMstack* stack=(EXE)->stack;\
-uint32_t ap=0;\
+size_t ap=0;\
 fklNiBegin(&ap,stack);
 #define FKL_NI_CHECK_TYPE(V,P,ERR_INFO,RUNNABLE,EXE) if(!P(V))FKL_RAISE_BUILTIN_ERROR_CSTR(ERR_INFO,FKL_ERR_WRONGARG,RUNNABLE,EXE)
 #ifdef __cplusplus
